fix: [GET] user list by query, not only employers
