+++ app/urls.py
@@ -5,7 +5,6 @@ from . import views
 urlpatterns = format_suffix_patterns([
     path('user', views.UserResponse.as_view()),
     path('user/token', views.UserTokenResponse.as_view()),
-    path('user/employers', views.UserEmployerResponse.as_view()),
     path('cover-letter', views.CoverLetterResponse.as_view()),
     path('cover-letter/<int:cl_id>', views.CoverLetterIDResponse.as_view()),
     path('position', views.PositionResponse.as_view()),
