+++ app/views/__init__.py
@@ -1,5 +1,4 @@
 from .user_response                 import *
 from .user_token_response           import *
-from .user_employer_response        import *
 from .cover_letter_response         import *
 from .position_response             import *
